Skilte seg BehandlingService i 2 deler. OpprettBehandlingService for opprettelse delen og BehandlingService for andre forbruk. Alle andre tjenester kommer til å kalle BehandlingService.kt(bortsett fra opprettelse flyt)
