+++ src/main/kotlin/no/nav/familie/ks/sak/api/BehandlingController.kt
@@ -7,6 +7,7 @@ import no.nav.familie.ks.sak.api.dto.OpprettBehandlingDto
 import no.nav.familie.ks.sak.common.exception.FunksjonellFeil
 import no.nav.familie.ks.sak.config.BehandlerRolle
 import no.nav.familie.ks.sak.kjerne.behandling.BehandlingService
+import no.nav.familie.ks.sak.kjerne.behandling.OpprettBehandlingService
 import no.nav.familie.ks.sak.sikkerhet.AuditLoggerEvent
 import no.nav.familie.ks.sak.sikkerhet.TilgangService
 import no.nav.security.token.support.core.api.ProtectedWithClaims
@@ -26,6 +27,7 @@ import org.springframework.web.bind.annotation.RestController
 @ProtectedWithClaims(issuer = "azuread")
 @Validated
 class BehandlingController(
+    private val opprettBehandlingService: OpprettBehandlingService,
     private val behandlingService: BehandlingService,
     private val tilgangService: TilgangService
 ) {
@@ -38,7 +40,7 @@ class BehandlingController(
             event = AuditLoggerEvent.CREATE,
             handling = "Opprett behandling"
         )
-        val behandling = behandlingService.opprettBehandling(opprettBehandlingDto)
+        val behandling = opprettBehandlingService.opprettBehandling(opprettBehandlingDto)
         return ResponseEntity.ok(Ressurs.success(behandlingService.lagBehandlingRespons(behandlingId = behandling.id)))
     }
